requirements.txt
aiohttp==3.8.1
aiorun==2022.4.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
python-engineio==4.3.4
python-socketio==5.7.1
yarl==1.8.1
